Split integration test code further
