+++ tests/test_integration_multihost_types.py
@@ -16,158 +16,8 @@ from .test_integration import (IntegrationTests, kpet_run_generate,
                                COMMONTREE_XML, create_asset_files)
 
 
-class IntegrationMultihostTests(IntegrationTests):
-    """Multihost integration tests"""
-
-    def test_multihost_no_types_no_regex_no_suites(self):
-        """Test multihost support without types/regex/suites"""
-        assets = {
-            "index.yaml": """
-                host_type_regex: ^normal
-                host_types:
-                    normal: {}
-                    panicky:
-                        ignore_panic: true
-                    multihost_1: {}
-                recipesets:
-                    rcs1:
-                      - normal
-                      - panicky
-                    rcs2:
-                      - multihost_1
-                      - multihost_2
-
-                arches:
-                    - arch
-                trees:
-                    tree: tree.xml
-            """,
-            "tree.xml": COMMONTREE_XML,
-        }
-
-        assets_path = create_asset_files(self, assets)
-
-        self.assertKpetSrcMatchesNoneOfTwoSuites(
-            assets_path)
-
-    def test_multihost_no_types_no_regex_two_suites(self):
-        """Test multihost support without types/regex and two suites"""
-        assets = {
-            "index.yaml": """
-                host_type_regex: ^normal
-                host_types:
-                    normal: {}
-                    panicky:
-                        ignore_panic: true
-                    multihost_1: {}
-                recipesets:
-                    rcs1:
-                      - normal
-                      - panicky
-                    rcs2:
-                      - multihost_1
-                      - multihost_2
-
-                arches:
-                    - arch
-                trees:
-                    tree: tree.xml
-                suites:
-                    - suite1.yaml
-                    - suite2.yaml
-            """,
-            "suite1.yaml": """
-                description: suite1
-                maintainers:
-                  - maint1
-                cases:
-                    - name: case1
-                      max_duration_seconds: 600
-                      pattern:
-                        sources:
-                          or:
-                            - a
-            """,
-            "suite2.yaml": """
-                description: suite2
-                maintainers:
-                  - maint1
-                cases:
-                    - name: case2
-                      max_duration_seconds: 600
-                      pattern:
-                        sources:
-                          or:
-                            - d
-            """,
-            "tree.xml": COMMONTREE_XML,
-        }
-
-        assets_path = create_asset_files(self, assets)
-
-        self.assertKpetSrcMatchesTwoSuites(
-            assets_path)
-
-    def test_multihost_no_types_wildcard_regex_two_suites(self):
-        """
-        Test multihost support without types, with a DB-level wildcard regex,
-        and two suites.
-        """
-        assets = {
-            "index.yaml": """
-                host_type_regex: ^normal
-                host_types:
-                    normal: {}
-                    panicky:
-                        ignore_panic: true
-                    multihost_1: {}
-                recipesets:
-                    rcs1:
-                      - normal
-                      - panicky
-                    rcs2:
-                      - multihost_1
-                      - multihost_2
-
-                arches:
-                    - arch
-                trees:
-                    tree: tree.xml
-                host_type_regex: .*
-                suites:
-                    - suite1.yaml
-                    - suite2.yaml
-            """,
-            "suite1.yaml": """
-                description: suite1
-                maintainers:
-                  - maint1
-                cases:
-                    - name: case1
-                      max_duration_seconds: 600
-                      pattern:
-                        sources:
-                          or:
-                            - a
-            """,
-            "suite2.yaml": """
-                description: suite2
-                maintainers:
-                  - maint1
-                cases:
-                    - name: case2
-                      max_duration_seconds: 600
-                      pattern:
-                        sources:
-                          or:
-                            - d
-            """,
-            "tree.xml": COMMONTREE_XML,
-        }
-
-        assets_path = create_asset_files(self, assets)
-
-        self.assertKpetSrcMatchesTwoSuites(assets_path)
+class IntegrationMultihostTypesTests(IntegrationTests):
+    """Multihost integration tests with at least one type"""
 
     def test_multihost_one_type_no_regex_no_suites(self):
         """Test multihost support with one type, but no regex/suites"""
